Fix orderlist resetting when clicking on person card
