+++ src/main/java/seedu/address/logic/Logic.java
@@ -51,4 +51,6 @@ public interface Logic {
      * Set the user prefs' GUI settings.
      */
     void setGuiSettings(GuiSettings guiSettings);
+
+    void clearOrderFilter();
 }
