+++ src/main/java/seedu/address/logic/LogicManager.java
@@ -91,4 +91,9 @@ public class LogicManager implements Logic {
     public void setGuiSettings(GuiSettings guiSettings) {
         model.setGuiSettings(guiSettings);
     }
+
+    @Override
+    public void clearOrderFilter() {
+        model.clearOrderFilter();
+    }
 }
