+++ src/main/java/seedu/address/model/Model.java
@@ -142,4 +142,6 @@ public interface Model {
      * @return the number of orders in the order list
      */
     int getOrderListSize();
+
+    void clearOrderFilter();
 }
