+++ src/main/java/seedu/address/ui/MainWindow.java
@@ -19,6 +19,7 @@ import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.order.Order;
+import seedu.address.model.order.OrderList;
 import seedu.address.model.person.Person;
 
 /**
@@ -58,6 +59,8 @@ public class MainWindow extends UiPart<Stage> {
     @FXML
     private StackPane statusbarPlaceholder;
 
+    private ObservableList<Order> allOrders;
+
     /**
      * Creates a {@code MainWindow} with the given {@code Stage} and {@code Logic}.
      */
@@ -74,6 +77,8 @@ public class MainWindow extends UiPart<Stage> {
         setAccelerators();
 
         helpWindow = new HelpWindow();
+
+        allOrders = logic.getFilteredOrderList();
     }
 
     public Stage getPrimaryStage() {
@@ -176,9 +181,6 @@ public class MainWindow extends UiPart<Stage> {
         primaryStage.hide();
     }
 
-    public PersonListPanel getPersonListPanel() {
-        return personListPanel;
-    }
 
     /**
      * Executes the command and returns the result.
@@ -186,6 +188,7 @@ public class MainWindow extends UiPart<Stage> {
      * @see seedu.address.logic.Logic#execute(String)
      */
     private CommandResult executeCommand(String commandText) throws CommandException, ParseException {
+        showAllOrders();
         try {
             CommandResult commandResult = logic.execute(commandText);
             logger.info("Result: " + commandResult.getFeedbackToUser());
@@ -215,15 +218,26 @@ public class MainWindow extends UiPart<Stage> {
      * @param person The {@link Person} whose orders are to be displayed. This object must not be {@code null}.
      */
     public void showPersonOrders(Person person) {
-        if (person.getOrders().isEmpty()) {
+        logic.clearOrderFilter();
+        if (person == null || person.getOrders().isEmpty()) {
             // Handle the case where there are no orders.
-            orderListPanel = new OrderListPanel(FXCollections.observableArrayList());
-            orderListPanelPlaceholder.getChildren().setAll(orderListPanel.getRoot());
+            orderListPanel.updateDisplayedOrders(FXCollections.observableArrayList());
+
         } else {
             // The person has orders; display them.
-            ObservableList<Order> orders = FXCollections.observableArrayList(person.getOrders());
-            orderListPanel = new OrderListPanel(orders);
-            orderListPanelPlaceholder.getChildren().setAll(orderListPanel.getRoot());
+            ObservableList<Order> filteredOrders = FXCollections.observableArrayList();
+            for (Order order : allOrders) {
+                if (person.getOrders().contains(order)) {
+                    filteredOrders.add(order);
+                }
+            }
+            orderListPanel.updateDisplayedOrders(filteredOrders);
+        }
+    }
+
+    public void showAllOrders() {
+        if (!allOrders.isEmpty()) {
+            orderListPanel.updateDisplayedOrders(allOrders);
         }
     }
 }
