+++ src/test/java/seedu/address/testutil/PersonBuilder.java
@@ -9,7 +9,7 @@ import seedu.address.model.person.Name;
 import seedu.address.model.person.Person;
 import seedu.address.model.person.Phone;
 import seedu.address.model.tag.Tag;
-import seedu.address.model.util.SampleDataUtil;
+import seedu.address.model.order.util.SampleDataUtil;
 
 /**
  * A utility class to help with building Person objects.
